sesion26/requirements.txt
django
djangorestframework
culqi